Move AthenzRoleResolver to upper level
